+++ include/bss-util/ArraySort.h
@@ -33,7 +33,7 @@ namespace bss {
     inline void Clear() { _array.Clear(); }
     inline void Discard(CType num) { _array.SetLength((num > _array.Length()) ? 0 : (_array.Length() - num)); }
     BSS_FORCEINLINE bool Empty() const { return _array.Empty(); }
-    BSS_FORCEINLINE void Reserve(CT_ capacity) { _array.Reserve(capacity); }
+    BSS_FORCEINLINE void SetCapacity(CT_ capacity) { _array.SetCapacity(capacity); }
     BSS_FORCEINLINE CT_ Length() const { return _array.Length(); }
     BSS_FORCEINLINE CT_ Capacity() const { return _array.Capacity(); }
     inline const T& Front() const { return _array.Front(); }